+++ cmd/imagefilter/main.go
@@ -1,0 +1,14 @@
+package main
+
+import (
+	"image-filter-editor/internal/gui"
+
+	"fyne.io/fyne/v2/app"
+)
+
+func main() {
+    a := app.NewWithID("computer-graphics.imagefilter")
+    window := gui.NewMainWindow(a)
+    window.Show()
+    a.Run()
+}
